Tutor-connect-system - AI/backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
stripe
channels
daphne
psycopg2-binary
python-dotenv
pandas
numpy
scikit-learn
joblib

langchain
langchain-community
langchain-openai

faiss-cpu
openai